proto/requirements.txt
fastapi
uvicorn
python-dotenv
imapclient
pydantic
openai